+++ tests/metrics/test_popularity.py
@@ -12,14 +12,15 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
+import numpy as np
 import pandas as pd
 import pytest
 
 from rectools import Columns
-from rectools.metrics.popularity import ARP
+from rectools.metrics.popularity import AvgRecPopularity
 
 
-class TestARP:
+class TestAvgRecPopularity:
     @pytest.fixture
     def interactions(self) -> pd.DataFrame:
         interactions = pd.DataFrame(
@@ -54,10 +55,52 @@ class TestARP:
     def test_correct_arp_values(
         self, recommendations: pd.DataFrame, interactions: pd.DataFrame, k: int, expected: pd.Series
     ) -> None:
-        arp = ARP(k)
+        arp = AvgRecPopularity(k)
 
         actual = arp.calc_per_user(recommendations, interactions)
         pd.testing.assert_series_equal(actual, expected, check_names=False)
 
         actual_mean = arp.calc(recommendations, interactions)
+        assert actual_mean == expected.mean()
+
+    def test_when_no_interactions(
+        self,
+        recommendations: pd.DataFrame,
+    ) -> None:
+        expected = pd.Series(index=recommendations[Columns.User].unique(), data=[0.0, 0.0, 0.0])
+        empty_interactions = pd.DataFrame(columns=[Columns.User, Columns.Item], dtype=int)
+        arp = AvgRecPopularity(k=2)
+
+        actual = arp.calc_per_user(recommendations, empty_interactions)
+        pd.testing.assert_series_equal(actual, expected, check_names=False)
+
+        actual_mean = arp.calc(recommendations, empty_interactions)
+        assert actual_mean == expected.mean()
+
+    @pytest.mark.parametrize(
+        "k,expected",
+        (
+            (1, pd.Series(index=["u1", "u2", "u3"], data=[3.0, 1.0, 1.0])),
+            (3, pd.Series(index=["u1", "u2", "u3"], data=[2.5, np.divide(4, 3), 1.5])),
+        ),
+    )
+    def test_when_new_item_in_reco(self, interactions: pd.DataFrame, k: int, expected: pd.Series) -> None:
+        reco = pd.DataFrame(
+            [
+                ["u1", "i1", 1],
+                ["u1", "i2", 2],
+                ["u2", "i3", 1],
+                ["u2", "i1", 2],
+                ["u2", "i4", 3],
+                ["u3", "i3", 1],
+                ["u3", "i2", 2],
+            ],
+            columns=[Columns.User, Columns.Item, Columns.Rank],
+        )
+        arp = AvgRecPopularity(k)
+
+        actual = arp.calc_per_user(reco, interactions)
+        pd.testing.assert_series_equal(actual, expected, check_names=False)
+
+        actual_mean = arp.calc(reco, interactions)
         assert actual_mean == expected.mean()